backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
python-docx
sentence-transformers
scikit-learn
torch
